MKT-544: Add BarChart component
