+++ package.json
@@ -21,7 +21,6 @@
     "@types/react": "^18.3.10",
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.2",
-    "add": "^2.0.6",
     "antd": "^5.21.2",
     "package-lock.json": "^1.0.0",
     "react": "^18.3.1",
