+++ src/appTheme.ts
@@ -1,0 +1,268 @@
+import { ThemeConfig } from 'antd'
+
+const j2Blue = {
+  1: 'var(--j2-blue-1)',
+  2: 'var(--j2-blue-2)',
+  3: 'var(--j2-blue-3)',
+  4: 'var(--j2-blue-4)',
+  5: 'var(--j2-blue-5)',
+  6: 'var(--j2-blue-6)',
+  7: 'var(--j2-blue-7)',
+  8: 'var(--j2-blue-8)',
+  9: 'var(--j2-blue-9)',
+  10: 'var(--j2-blue-10)',
+  11: 'var(--j2-blue-11)',
+  12: 'var(--j2-blue-12)',
+}
+
+const j2Gray = {
+  1: 'var(--j2-gray-1)',
+  2: 'var(--j2-gray-2)',
+  3: 'var(--j2-gray-3)',
+  4: 'var(--j2-gray-4)',
+  5: 'var(--j2-gray-5)',
+  6: 'var(--j2-gray-6)',
+  7: 'var(--j2-gray-7)',
+  8: 'var(--j2-gray-8)',
+  9: 'var(--j2-gray-9)',
+  10: 'var(--j2-gray-10)',
+  11: 'var(--j2-gray-11)',
+  12: 'var(--j2-gray-12)',
+}
+
+const j2Green = {
+  1: 'var(--j2-green-1)',
+  2: 'var(--j2-green-2)',
+  3: 'var(--j2-green-3)',
+  4: 'var(--j2-green-4)',
+  5: 'var(--j2-green-5)',
+  6: 'var(--j2-green-6)',
+  7: 'var(--j2-green-7)',
+  8: 'var(--j2-green-8)',
+  9: 'var(--j2-green-9)',
+  10: 'var(--j2-green-10)',
+  11: 'var(--j2-green-11)',
+  12: 'var(--j2-green-12)',
+}
+
+const j2Gold = {
+  1: 'var(--j2-gold-1)',
+  2: 'var(--j2-gold-2)',
+  3: 'var(--j2-gold-3)',
+  4: 'var(--j2-gold-4)',
+  5: 'var(--j2-gold-5)',
+  6: 'var(--j2-gold-6)',
+  7: 'var(--j2-gold-7)',
+  8: 'var(--j2-gold-8)',
+  9: 'var(--j2-gold-9)',
+  10: 'var(--j2-gold-10)',
+  11: 'var(--j2-gold-11)',
+  12: 'var(--j2-gold-12)',
+}
+
+const j2Red = {
+  1: 'var(--j2-red-1)',
+  2: 'var(--j2-red-2)',
+  3: 'var(--j2-red-3)',
+  4: 'var(--j2-red-4)',
+  5: 'var(--j2-red-5)',
+  6: 'var(--j2-red-6)',
+  7: 'var(--j2-red-7)',
+  8: 'var(--j2-red-8)',
+  9: 'var(--j2-red-9)',
+  10: 'var(--j2-red-10)',
+  11: 'var(--j2-red-11)',
+  12: 'var(--j2-red-12)',
+}
+
+export const appTheme: ThemeConfig = {
+  token: {
+    // Colors
+    colorText: j2Gray[12],
+    colorTextSecondary: j2Gray[11],
+    colorTextTertiary: j2Gray[11],
+    colorTextQuaternary: j2Gray[10],
+    colorTextHeading: j2Gray[12],
+    colorTextLabel: j2Gray[11],
+    colorTextDescription: j2Gray[11],
+    colorTextDisabled: j2Gray[10],
+    colorTextPlaceholder: j2Gray[10],
+    colorIcon: j2Gray[11],
+    colorIconHover: j2Gray[12],
+    colorBgLayout: j2Gray[2],
+    colorBorder: j2Gray[7],
+    colorBorderSecondary: j2Gray[6],
+    colorPrimary: j2Blue[9],
+    colorSuccess: j2Green[9],
+    colorWarning: j2Gold[9],
+    colorInfo: j2Blue[9],
+    colorError: j2Red[9],
+    colorLink: j2Blue[9],
+    colorErrorBg: j2Red[2],
+    colorErrorBgHover: j2Red[3],
+    colorErrorBorder: j2Red[7],
+    colorErrorBorderHover: j2Red[8],
+    colorErrorHover: j2Red[10],
+    colorErrorActive: j2Red[10],
+    colorErrorTextHover: j2Red[12],
+    colorErrorText: j2Red[11],
+    colorErrorTextActive: j2Red[12],
+    colorLinkHover: j2Blue[10],
+    colorInfoBg: j2Blue[1],
+    colorInfoBgHover: j2Blue[4],
+    colorInfoBorder: j2Blue[7],
+    colorInfoBorderHover: j2Blue[8],
+    colorInfoHover: j2Blue[10],
+    colorInfoActive: j2Blue[10],
+    colorInfoTextHover: j2Blue[11],
+    colorInfoText: j2Blue[12],
+    colorInfoTextActive: j2Blue[11],
+    colorLinkActive: j2Blue[10],
+    colorPrimaryBg: j2Blue[1],
+    colorPrimaryBgHover: j2Blue[4],
+    colorPrimaryBorder: j2Blue[7],
+    colorPrimaryBorderHover: j2Blue[8],
+    colorPrimaryHover: j2Blue[10],
+    colorPrimaryActive: j2Blue[10],
+    colorPrimaryTextHover: j2Blue[11],
+    colorPrimaryText: j2Blue[12],
+    colorPrimaryTextActive: j2Blue[11],
+    colorSuccessBg: j2Green[2],
+    colorSuccessBgHover: j2Green[3],
+    colorSuccessBorder: j2Green[7],
+    colorSuccessBorderHover: j2Green[8],
+    colorSuccessHover: j2Green[10],
+    colorSuccessActive: j2Green[10],
+    colorSuccessTextHover: j2Green[12],
+    colorSuccessText: j2Green[11],
+    colorSuccessTextActive: j2Green[12],
+    colorWarningBg: j2Gold[2],
+    colorWarningBgHover: j2Gold[3],
+    colorWarningBorder: j2Gold[7],
+    colorWarningBorderHover: j2Gold[8],
+    colorWarningHover: j2Gold[10],
+    colorWarningActive: j2Gold[10],
+    colorWarningTextHover: j2Gold[12],
+    colorWarningText: j2Gold[11],
+    colorWarningTextActive: j2Gold[12],
+    controlItemBgActive: j2Blue[1],
+    controlItemBgActiveHover: j2Blue[4],
+    colorFill: 'rgba(0, 0, 0, 0.15)',
+    colorFillSecondary: 'rgba(0, 0, 0, 0.06)',
+    colorFillTertiary: 'rgba(0, 0, 0, 0.04)',
+    colorFillQuaternary: 'rgba(0, 0, 0, 0.02)',
+    colorFillContent: 'rgba(0, 0, 0, 0.06)',
+    colorFillContentHover: 'rgba(0, 0, 0, 0.15)',
+    colorBgMask: 'rgba(32, 32, 27, 0.45)',
+    colorBgSpotlight: 'rgba(32, 32, 27, 0.85)',
+    colorBgContainerDisabled: 'rgba(0, 0, 0, 0.04)',
+    colorBgTextActive: 'rgba(0, 0, 0, 0.15)',
+    colorBgTextHover: 'rgba(0, 0, 0, 0.06)',
+    colorFillAlter: 'rgba(0, 0, 0, 0.02)',
+    colorSplit: 'rgba(0, 0, 0, 0.06)',
+    colorErrorOutline: 'rgba(255, 38, 6, 0.06)',
+    colorWarningOutline: 'rgba(255, 215, 5, 0.1)',
+    controlItemBgActiveDisabled: 'rgba(0, 0, 0, 0.15)',
+    controlItemBgHover: 'rgba(0, 0, 0, 0.04)',
+    controlOutline: 'rgba(37, 55, 97, 0.1)',
+    controlTmpOutline: 'rgba(0, 0, 0, 0.02)',
+
+    // Typography
+    fontFamily: 'Rubik, sans-serif',
+    fontFamilyCode: 'Rubik, sans-serif',
+    fontSize: 14,
+    fontSizeLG: 16,
+    fontSizeSM: 12,
+    fontSizeXL: 20,
+    fontSizeHeading1: 38,
+    fontSizeHeading2: 30,
+    fontSizeHeading3: 24,
+    fontSizeHeading4: 20,
+    fontSizeHeading5: 16,
+    lineHeight: 1.5714285714285714,
+    lineHeightHeading1: 1.2105263157894737,
+    lineHeightHeading2: 1.2666666666666666,
+    lineHeightHeading3: 1.3333333333333333,
+    lineHeightHeading4: 1.4,
+    lineHeightHeading5: 1.5,
+    lineHeightLG: 1.5,
+    lineHeightSM: 1.6666666666666667,
+    fontSizeIcon: 12,
+    fontWeightStrong: 600,
+
+    // Border
+    borderRadius: 8,
+    borderRadiusLG: 10,
+    borderRadiusSM: 6,
+    borderRadiusXS: 2,
+
+    // Spacing and sizing
+    sizeStep: 4,
+    sizeUnit: 4,
+    controlInteractiveSize: 16,
+    size: 16,
+    sizeLG: 24,
+    sizeMD: 20,
+    sizeMS: 16,
+    sizeSM: 12,
+    sizeXL: 32,
+    sizeXS: 8,
+    sizeXXL: 48,
+    controlHeight: 32,
+    sizeXXS: 4,
+    controlHeightLG: 40,
+    controlHeightSM: 24,
+    controlHeightXS: 16,
+    lineWidth: 1,
+    lineWidthBold: 2,
+    lineWidthFocus: 4,
+    controlOutlineWidth: 2,
+    screenLG: 992,
+    screenLGMax: 1199,
+    screenLGMin: 992,
+    screenMD: 768,
+    screenMDMax: 991,
+    screenMDMin: 768,
+    screenSM: 576,
+    screenSMMax: 767,
+    screenSMMin: 576,
+    screenXL: 1200,
+    screenXLMax: 1599,
+    screenXLMin: 1200,
+    screenXS: 480,
+    screenXSMax: 575,
+    screenXSMin: 480,
+    screenXXL: 1600,
+    screenXXLMin: 1600,
+    sizePopupArrow: 16,
+    margin: 16,
+    marginLG: 24,
+    marginMD: 20,
+    marginSM: 12,
+    marginXL: 32,
+    marginXS: 8,
+    marginXXL: 48,
+    marginXXS: 4,
+    padding: 16,
+    paddingLG: 24,
+    paddingMD: 20,
+    paddingSM: 12,
+    paddingXL: 32,
+    paddingXS: 8,
+    paddingXXS: 4,
+    paddingContentHorizontal: 16,
+    paddingContentHorizontalLG: 24,
+    paddingContentHorizontalSM: 16,
+    paddingContentVertical: 12,
+    paddingContentVerticalLG: 16,
+    paddingContentVerticalSM: 8,
+    controlPaddingHorizontal: 12,
+    controlPaddingHorizontalSM: 8,
+  },
+  components: {
+    Menu: {
+      itemSelectedBg: j2Blue[10],
+      itemSelectedColor: '#fff',
+    },
+  },
+}
